requirements.txt
bedrock-agentcore>=1.0.0
strands-agents>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
boto3>=1.28.0
matplotlib>=3.7.0
